backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
langchain>=0.3.0
langgraph>=0.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
supabase>=2.0.0
httpx>=0.28.0
